Modify Pokerbot class to IrcBot
